+++ apk/Changelog.md
@@ -9,6 +9,15 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Version December 29 2024 🎉 🎊 🥂 ✨ 🎅 💖 🤶 ✨ 🥂 🎊 🎉
+==============
+
+* Add support for Shared chat
+* Add new settings option Chat > Disable shared chat (No by default)
+* Add new settings option Chat > Show Shared chat badges (Yes by default)
+* Add new Player controls Chat extra settings > Disable shared chat (No by default)
+* 🎉 🎊 🥂 ✨ 🎅 💖 🤶 ✨ 🥂 🎊 🎉 Happy New Year, Merry Christmas, and Happy Holidays everybody!!! 🎉 🎊 🥂 ✨ 🎅 💖 🤶 ✨ 🥂 🎊 🎉
+
 Version December 28 2024 🎉 🎊 🥂 ✨ 🎅 💖 🤶 ✨ 🥂 🎊 🎉
 ==============